Moved tag functions into new file
